add beforeAll, afterAll, beforeEach, afterEach
hooks.

Due to the pattern matching the code is manual and messy. It also comes with the limitation of using a limited combination of these hooks in a predefined order. I will look into changing that in the future, but still this is better than not having a hook capability